requirements.txt
scrapy
psycopg2-binary
python-dotenv
apache-airflow
apache-airflow-providers-docker
apache-airflow-providers-postgres
sqlalchemy
requests

